requirements.txt
# Web Framework 
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12     

# Data Validation
pydantic==2.10.3

# Document Processing 
python-docx==1.1.2
pymupdf           

# Embeddings 
sentence-transformers==3.3.1 
torch==2.5.1                

# Vector Math (Cosine Similarity) ---
numpy==2.1.3                 

# Local LLM
llama-cpp-python

# Development & Testing
httpx==0.28.1              
pytest==8.3.4                
pytest-asyncio==0.24.0      